Coffee Shop
was in dire need of a coffee, so before taking a study break, I developed a interact coffee shop barrista

With some variables & conditions having a different interaction than others (sorry Dr Doom) I wanted to have a script that might hep a starting barista engage in conversation/develop a routine they can use when working
(ok, so i can't make the coffee for them with the script)
but this will give a routine for them to take, improve and develop into their own to improve their customer service

I also have laid the field, as I am buying a coffee shop before my 34th birthday, where staff of all abilities and varied diversities can work, to engage with customer and improve their skills, as well as delivering incredible coffee! (and cakes, gotta have a delightful cookie or two with a mean vanilla latte)

I'll come back here and there to expand the code, adding further scenarios, but for now, i'm uploading and off to get myself a chai latte!